--- src/__init__.py
@@ -1,1 +1,0 @@
-import os, sys; sys.path.append(os.path.dirname(os.path.realpath(__file__)))
