+++ talklib.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+LICENSE.txt
+README.md
+pyproject.toml
+talklib/__init__.py
+talklib/ev.py
+talklib/show.py
+talklib/utils.py
+talklib.egg-info/PKG-INFO
+talklib.egg-info/SOURCES.txt
+talklib.egg-info/dependency_links.txt
+talklib.egg-info/requires.txt
+talklib.egg-info/top_level.txt
