+++ talklib.egg-info/requires.txt
@@ -1,0 +1,21 @@
+attrs>=22.1.0
+certifi>=2022.9.24
+charset-normalizer>=2.1.1
+colorama>=0.4.6
+coverage>=6.5.0
+exceptiongroup>=1.0.4
+ffmpeg-python>=0.2.0
+future>=0.18.3
+idna>=3.4
+iniconfig>=1.1.1
+packaging>=21.3
+pluggy>=1.0.0
+PyJWT>=2.6.0
+pyparsing>=3.0.9
+pytest>=7.2.0
+pytest-cov>=4.0.0
+pytz>=2022.6
+requests>=2.28.1
+tomli>=2.0.1
+twilio>=7.15.3
+urllib3>=1.26.13
