+++ talklib/__init__.py
@@ -1,4 +1,4 @@
 import os, sys; sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 
-from .src.show import TLShow
-from .src import utils
+from show import TLShow
+import utils
