+++ talklib/ev.py
