+++ talklib/show.py
@@ -22,8 +22,8 @@ from datetime import datetime
 import ffmpeg
 import requests
 
-from .ev import EV
-from .utils import get_timestamp, clear_screen, print_to_screen, today_is_weekday, send_sms
+from talklib.ev import EV
+from talklib.utils import get_timestamp, clear_screen, print_to_screen, today_is_weekday, send_sms
 
 cwd = os.getcwd()
